requirements.txt
coverage==5.0.3
Django==2.2
django-cors-middleware==1.5.0
django-filter==2.2.0
django-redis==4.11.0
djangorestframework==3.11.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
hiredis==1.0.1
Markdown==3.1.1
psycopg2-binary==2.8.4
python-rapidjson==0.9.1
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.0
